requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
tabulate>=0.9.0
pytest>=7.0.0
streamlit>=1.30.0
